+++ src/main/java/entities/DynamicEvent.java
@@ -9,9 +9,7 @@ import java.sql.Time;
 import java.time.LocalDateTime;
 
 
-public class DynamicEvent extends Event{
-
-    private String dynamicGoal;
+public class DynamicEvent implements Event{
 
     /** Constructs a DynamicEvent with a name, startTime, endTime, commuteTime, and location.
      *
@@ -21,16 +19,54 @@ public class DynamicEvent extends Event{
      * @param commute the time it takes to get to this event
      * @param location the location of the event to be created
      */
-    public DynamicEvent(String name, LocalDateTime startTime, LocalDateTime endTime, Boolean isCommute, Event commute, String location){
-        super(name, startTime, endTime, isCommute, commute, location);
+    private String name;
+    private LocalDateTime startTime;
+    private LocalDateTime endTime;
+    private boolean isCommute;
+    private String location;
+    private Event commute;
+
+
+    public DynamicEvent(String name, LocalDateTime startTime, LocalDateTime endTime, boolean isCommute, Event commute, String location){
+
+        this.name = name;
+        this.startTime = startTime;
+        this.endTime = endTime;
+        this.commute = commute;
+        this.location = location;
+        this.isCommute = isCommute;
+    }
+    @Override
+    public String getName() {
+        return this.name;
     }
 
-    public String getDynamicGoal(){
-        return this.dynamicGoal;
+    @Override
+    public Event getCommute() {
+        return this.commute;
+    }
+    @Override
+    public LocalDateTime getStartTime() {
+        return startTime;
+    }
+    @Override
+    public LocalDateTime getEndTime() {
+        return endTime;
+    }
+    @Override
+    public String getLocation() {
+        return location;
+    }
+    @Override
+    public boolean isCommute(){
+        return this.isCommute;
     }
 
-    public void setDynamicGoal(String dynamicGoal){
-        this.dynamicGoal = dynamicGoal;
+    @Override
+    public boolean isValid() {
+        if (endTime.isBefore(startTime)) {
+            return false;
+        }
+        return true;
     }
-
 }
