+++ src/main/java/entities/SingleEvent.java
@@ -9,7 +9,7 @@ import java.sql.Time;
 import java.time.LocalDateTime;
 
 
-public class SingleEvent extends Event {
+public class SingleEvent implements Event {
 
 
     /**
@@ -21,7 +21,54 @@ public class SingleEvent extends Event {
      * @param commute commute time to get to the event (optional, may be null)
      * @param location the location of the event (optional, may be null)
      */
-    public SingleEvent(String name, LocalDateTime startTime, LocalDateTime endTime, Boolean isCommute, Event commute, String location){
-        super(name, startTime, endTime, isCommute, commute, location);
+    private String name;
+    private LocalDateTime startTime;
+    private LocalDateTime endTime;
+    private boolean isCommute;
+    private String location;
+    private Event commute;
+
+
+    public SingleEvent(String name, LocalDateTime startTime, LocalDateTime endTime, boolean isCommute, Event commute, String location){
+
+        this.name = name;
+        this.startTime = startTime;
+        this.endTime = endTime;
+        this.commute = commute;
+        this.location = location;
+        this.isCommute = isCommute;
+    }
+    @Override
+    public String getName() {
+        return this.name;
+    }
+
+    @Override
+    public Event getCommute() {
+        return this.commute;
+    }
+    @Override
+    public LocalDateTime getStartTime() {
+        return startTime;
+    }
+    @Override
+    public LocalDateTime getEndTime() {
+        return endTime;
+    }
+    @Override
+    public String getLocation() {
+        return location;
+    }
+    @Override
+    public boolean isCommute(){
+        return this.isCommute;
+    }
+
+    @Override
+    public boolean isValid() {
+        if (endTime.isBefore(startTime)) {
+            return false;
+        }
+        return true;
     }
 }
